Moved data observatory collections to data.
Updated references.
Updated specs.
