+++ lib/assets/core/javascripts/cartodb3/components/form-components/editors/data-observatory/dropdown-view.js
@@ -3,8 +3,8 @@ var template = require('./dropdown.tpl');
 var selectedItemTemplate = require('./dropdown-item.tpl');
 var DropdownDialogView = require('./dropdown-dialog-view');
 var PopupManager = require('../../../popup-manager');
-var MeasurementsCollection = require('./measurements-collection');
-var FiltersCollection = require('./measurements-filters-collection');
+var MeasurementsCollection = require('../../../../data/data-observatory/measurements-collection');
+var FiltersCollection = require('../../../../data/data-observatory/filters-collection');
 
 var ENTER_KEY_CODE = 13;
 var STATE = {
