+++ lib/assets/core/javascripts/cartodb3/data/data-observatory/data-observatory-base-collection.js
@@ -1,8 +1,8 @@
 var Backbone = require('backbone');
 var _ = require('underscore');
 var cdb = require('cartodb.js');
-var errorParse = require('../../../../helpers/error-parser');
-var checkAndBuildOpts = require('../../../../helpers/required-opts');
+var errorParse = require('../../helpers/error-parser');
+var checkAndBuildOpts = require('../../helpers/required-opts');
 
 var REQUIRED_OPTS = [
   'configModel',
@@ -28,11 +28,10 @@ module.exports = Backbone.Collection.extend({
   fetch: function (options) {
     this._success = options && options.success;
     this._error = options && options.error;
-    var table = this._layerDefinitionModel.getTableName();
+
     var query = this._querySchemaModel.get('query');
 
     var sqlQuery = this._queryTemplate({
-      table: table,
       query: query,
       country: this._country
     });
