+++ lib/assets/core/javascripts/cartodb3/data/data-observatory/filters-collection.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
-var BaseCollection = require('./measurements-base-collection');
-var BaseModel = require('../../../custom-list/custom-list-item-model');
+var BaseCollection = require('./data-observatory-base-collection');
+var BaseModel = require('../../components/custom-list/custom-list-item-model');
 
 var FILTERS_QUERY = "SELECT count(*) num_measurements, tag.key subsection_id, tag.value subsection_name FROM OBS_GetAvailableNumerators((SELECT ST_SetSRID(ST_Extent(the_geom), 4326) FROM (<%- query %>) q), <%- country %>) numers, Jsonb_Each(numers.numer_tags) tag WHERE tag.key like 'subsection%' GROUP BY tag.key, tag.value";
