+++ lib/assets/core/javascripts/cartodb3/data/data-observatory/measurements-collection.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
-var BaseCollection = require('./measurements-base-collection');
-var BaseModel = require('../../../custom-list/custom-list-item-model');
+var BaseCollection = require('./data-observatory-base-collection');
+var BaseModel = require('../../components/custom-list/custom-list-item-model');
 
 var MEASUREMENTS_QUERY = 'SELECT * FROM OBS_GetAvailableNumerators((SELECT ST_SetSRID(ST_Extent(the_geom), 4326) FROM (<%- query %>) q), <%- country %>) numers';
