+++ lib/assets/core/test/spec/cartodb3/components/form-components/editors/data-observatory/dropdown-dialog-view.spec.js
@@ -2,8 +2,8 @@ var Backbone = require('backbone');
 var cdb = require('cartodb.js');
 var DropdownDialogView = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/dropdown-dialog-view');
 var LayerDefinitionModel = require('../../../../../../../javascripts/cartodb3/data/layer-definition-model');
-var MeasurementsCollection = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/measurements-collection');
-var FiltersCollection = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/measurements-filters-collection');
+var MeasurementsCollection = require('../../../../../../../javascripts/cartodb3/data/data-observatory/measurements-collection');
+var FiltersCollection = require('../../../../../../../javascripts/cartodb3/data/data-observatory/filters-collection');
 
 describe('components/form-components/editors/data-observatory/dropdown-dialog-view', function () {
   beforeEach(function () {
