+++ lib/assets/core/test/spec/cartodb3/components/form-components/editors/data-observatory/filter-list-view.spec.js
@@ -2,7 +2,7 @@ var Backbone = require('backbone');
 var cdb = require('cartodb.js');
 
 var FilterListView = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/filter-list-view');
-var FiltersCollection = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/measurements-filters-collection');
+var FiltersCollection = require('../../../../../../../javascripts/cartodb3/data/data-observatory/filters-collection');
 var LayerDefinitionModel = require('../../../../../../../javascripts/cartodb3/data/layer-definition-model');
 
 describe('components/form-components/editors/data-observatory/filter-list-view', function () {
