+++ lib/assets/core/test/spec/cartodb3/components/form-components/editors/data-observatory/measurements-list-view.spec.js
@@ -1,9 +1,9 @@
 var Backbone = require('backbone');
 var cdb = require('cartodb.js');
 
-var MeasurementsCollection = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/measurements-collection');
+var MeasurementsCollection = require('../../../../../../../javascripts/cartodb3/data/data-observatory/measurements-collection');
 var MeasurementListView = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/measurement-list-view');
-var FiltersCollection = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/measurements-filters-collection');
+var FiltersCollection = require('../../../../../../../javascripts/cartodb3/data/data-observatory/filters-collection');
 var LayerDefinitionModel = require('../../../../../../../javascripts/cartodb3/data/layer-definition-model');
 
 describe('components/form-components/editors/data-observatory/measurement-list-view', function () {
