+++ lib/assets/core/test/spec/cartodb3/data/data-observatory/filters-collection.spec.js
@@ -1,10 +1,10 @@
 var Backbone = require('backbone');
 var _ = require('underscore');
-var FilterCollection = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/measurements-filters-collection');
-var LayerDefinitionModel = require('../../../../../../../javascripts/cartodb3/data/layer-definition-model');
+var FilterCollection = require('../../../../../javascripts/cartodb3/data/data-observatory/filters-collection');
+var LayerDefinitionModel = require('../../../../../javascripts/cartodb3/data/layer-definition-model');
 var cdb = require('cartodb.js');
 
-describe('components/form-components/editors/data-observatory/measurements-filters-collection', function () {
+describe('data/data-observatory/filters-collection', function () {
   beforeEach(function () {
     var configModel = new Backbone.Model({
       base_url: '/u/foo',
