+++ lib/assets/core/test/spec/cartodb3/data/data-observatory/measurements-collection.spec.js
@@ -1,9 +1,9 @@
 var Backbone = require('backbone');
 var cdb = require('cartodb.js');
-var MeasurementsCollection = require('../../../../../../../javascripts/cartodb3/components/form-components/editors/data-observatory/measurements-collection');
-var LayerDefinitionModel = require('../../../../../../../javascripts/cartodb3/data/layer-definition-model');
+var MeasurementsCollection = require('../../../../../javascripts/cartodb3/data/data-observatory/measurements-collection');
+var LayerDefinitionModel = require('../../../../../javascripts/cartodb3/data/layer-definition-model');
 
-describe('components/form-components/editors/data-observatory/measurements-collection', function () {
+describe('data/data-observatory/measurements-collection', function () {
   beforeEach(function () {
     var configModel = new Backbone.Model({
       base_url: '/u/foo',
